# contracts/host/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisNoteScope": {
        "anyOf": [
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "enum": [
                  "global"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "contextId": {
                "pattern": "^[1-9]\\d{0,15}$",
                "type": "string"
              },
              "kind": {
                "enum": [
                  "context"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind",
              "contextId"
            ],
            "type": "object"
          }
        ]
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "parameters": {
            "additionalProperties": false,
            "properties": {
              "currentRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              "expectedRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              }
            },
            "type": "object"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId",
          "retryable",
          "parameters"
        ],
        "type": "object"
      },
...
      "WorkScope": {
        "anyOf": [
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "enum": [
                  "free"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "contextId": {
                "pattern": "^[1-9]\\d{0,15}$",
                "type": "string"
              },
              "kind": {
                "enum": [
                  "context"
                ],
                "type": "string"
              }
            },
            "required": [
              "kind",
              "contextId"
            ],
            "type": "object"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "hostBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Plysmith Application Host",
    "version": "0.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/analysis/notes": {
      "post": {
        "operationId": "CreateAnalysisNote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "expectedScratchId": {
                    "maxLength": 160,
                    "minLength": 1,
                    "pattern": "^[A-Za-z0-9._:-]+$",
                    "type": "string"
                  },
                  "expectedScratchRevision": {
                    "maximum": 9007199254740991,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "languageTag": {
                    "maxLength": 35,
                    "minLength": 1,
                    "type": "string"
                  },
                  "noteScope": {
                    "$ref": "#/components/schemas/AnalysisNoteScope"
                  },
                  "scope": {
                    "$ref": "#/components/schemas/WorkScope"
                  }
                },
                "required": [
                  "scope",
                  "expectedScratchId",
                  "expectedScratchRevision",
                  "languageTag",
                  "noteScope"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnalysisNoteResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
...
    "/analysis/position-notes": {
      "post": {
        "operationId": "CreatePositionNote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "anchorId": {
                    "pattern": "^[1-9]\\d{0,15}$",
                    "type": "string"
                  },
                  "body": {
                    "maxLength": 8000,
                    "minLength": 1,
                    "type": "string"
                  },
                  "itemId": {
                    "pattern": "^[1-9]\\d{0,15}$",
                    "type": "string"
                  },
                  "languageTag": {
                    "maxLength": 35,
                    "minLength": 1,
                    "type": "string"
                  },
                  "noteScope": {
                    "$ref": "#/components/schemas/AnalysisNoteScope"
                  },
                  "revisionId": {
                    "pattern": "^[1-9]\\d{0,15}$",
                    "type": "string"
                  },
                  "scope": {
                    "$ref": "#/components/schemas/WorkScope"
                  }
                },
                "required": [
                  "scope",
                  "itemId",
                  "revisionId",
                  "anchorId",
                  "body",
                  "languageTag",
                  "noteScope"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisNoteMutationResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
    "/analysis/scratch": {
      "put": {
        "operationId": "UpdateAnalysisScratch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "action": {
                    "anyOf": [
                      {
                        "additionalProperties": false,
                        "properties": {
                          "kind": {
                            "enum": [
                              "start"
                            ],
                            "type": "string"
                          },
                          "origin": {
                            "anyOf": [
                              {
                                "additionalProperties": false,
                                "properties": {
                                  "kind": {
                                    "enum": [
                                      "initial_position"
                                    ],
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "kind"
                                ],
                                "type": "object"
                              },
                              {
                                "additionalProperties": false,
                                "properties": {
                                  "fen": {
                                    "maxLength": 128,
                                    "minLength": 1,
                                    "type": "string"
                                  },
                                  "kind": {
                                    "enum": [
                                      "fen"
                                    ],
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "kind",
                                  "fen"
                                ],
                                "type": "object"
                              },
                              {
                                "additionalProperties": false,
                                "properties": {
                                  "anchorId": {
                                    "pattern": "^[1-9]\\d{0,15}$",
                                    "type": "string"
                                  },
                                  "itemId": {
                                    "pattern": "^[1-9]\\d{0,15}$",
                                    "type": "string"
                                  },
                                  "kind": {
                                    "enum": [
                                      "inventory_anchor"
                                    ],
                                    "type": "string"
                                  },
                                  "revisionId": {
                                    "pattern": "^[1-9]\\d{0,15}$",
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "kind",
                                  "itemId",
                                  "revisionId",
                                  "anchorId"
                                ],
                                "type": "object"
                              }
                            ]
                          }
                        },
                        "required": [
                          "kind",
                          "origin"
                        ],
                        "type": "object"
                      },
                      {
                        "additionalProperties": false,
                        "properties": {
                          "kind": {
                            "enum": [
                              "apply_move"
                            ],
                            "type": "string"
                          },
                          "move": {
                            "anyOf": [
                              {
                                "additionalProperties": false,
                                "properties": {
                                  "kind": {
                                    "enum": [
                                      "coordinates"
                                    ],
                                    "type": "string"
                                  },
                                  "value": {
                                    "maxLength": 5,
                                    "minLength": 4,
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "kind",
                                  "value"
                                ],
                                "type": "object"
                              },
                              {
                                "additionalProperties": false,
                                "properties": {
                                  "kind": {
                                    "enum": [
                                      "notation"
                                    ],
                                    "type": "string"
                                  },
                                  "locale": {
                                    "anyOf": [
                                      {
                                        "enum": [
                                          "de-DE"
                                        ],
                                        "type": "string"
                                      },
                                      {
                                        "enum": [
                                          "en-GB"
                                        ],
                                        "type": "string"
                                      }
                                    ]
                                  },
                                  "value": {
                                    "maxLength": 32,
                                    "minLength": 1,
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "kind",
                                  "value",
                                  "locale"
                                ],
                                "type": "object"
                              }
                            ]
                          }
                        },
                        "required": [
                          "kind",
                          "move"
                        ],
                        "type": "object"
                      },
                      {
                        "additionalProperties": false,
                        "properties": {
                          "cursor": {
                            "maximum": 9007199254740991,
                            "minimum": 0,
                            "type": "integer"
                          },
                          "kind": {
                            "enum": [
                              "move_cursor"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "kind",
                          "cursor"
                        ],
                        "type": "object"
                      },
                      {
                        "additionalProperties": false,
                        "properties": {
                          "body": {
                            "maxLength": 8000,
                            "type": "string"
                          },
                          "kind": {
                            "enum": [
                              "prepare_note"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "kind",
                          "body"
                        ],
                        "type": "object"
                      },
                      {
                        "additionalProperties": false,
                        "properties": {
                          "kind": {
                            "enum": [
                              "clear_note"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "kind"
                        ],
                        "type": "object"
                      },
                      {
                        "additionalProperties": false,
                        "properties": {
                          "kind": {
                            "enum": [
                              "discard"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "kind"
                        ],
                        "type": "object"
                      }
                    ]
                  },
                  "expectedScratchId": {
                    "anyOf": [
                      {
                        "maxLength": 160,
                        "minLength": 1,
                        "pattern": "^[A-Za-z0-9._:-]+$",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "expectedScratchRevision": {
                    "anyOf": [
                      {
                        "maximum": 9007199254740991,
                        "minimum": 1,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "scope": {
                    "$ref": "#/components/schemas/WorkScope"
                  }
                },
                "required": [
                  "scope",
                  "expectedScratchId",
                  "expectedScratchRevision",
                  "action"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAnalysisScratchResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
...
    "/events": {
      "get": {
        "operationId": "SubscribeHostEvents",
        "parameters": [
          {
            "in": "header",
            "name": "last-event-id",
            "required": false,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "x-host-event-schema": {
                    "$ref": "#/components/schemas/HostEvent"
                  }
                }
              }
            },
            "description": "SSE frames use event: host-event, id: eventId and JSON data matching HostEvent. A host.replay-gap requires fresh query snapshots."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
...
  },
  "security": [
    {
      "hostBearer": []
    }
  ]
}